+++ src/InvalidArgumentException.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * @copyright ©2017 Quicken Loans Inc. All rights reserved. Trade Secret,
+ *    Confidential and Proprietary. Any dissemination outside of Quicken Loans
+ *    is strictly prohibited.
+ */
+namespace QL\MCP\Cache;
+
+class InvalidArgumentException extends Exception implements \Psr\SimpleCache\InvalidArgumentException
+{
+}