+++ src/MemoryCache.php
@@ -7,13 +7,18 @@
 
 namespace QL\MCP\Cache;
 
+use QL\MCP\Cache\CacheInterface as MCPCacheInterface;
+use Psr\SimpleCache\CacheInterface as PSR16CacheInterface;
 use QL\MCP\Cache\Item\Item;
+use QL\MCP\Cache\Utility\CacheInputValidationTrait;
 
 /**
  * @internal
  */
-class MemoryCache implements CacheInterface
+class MemoryCache implements PSR16CacheInterface, MCPCacheInterface
 {
+    use CacheInputValidationTrait;
+
     /**
      * @var Item[]
      */
@@ -25,25 +30,168 @@ class MemoryCache implements CacheInterface
     }
 
     /**
-     * {@inheritdoc}
+     * Fetches a value from the cache.
+     *
+     * @param string $key The unique key of this item in the cache.
+     * @param mixed $default Default value to return if the key does not exist.
+     *
+     * @return mixed The value of the item from the cache, or $default in case of cache miss.
+     *
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     *   MUST be thrown if the $key string is not a legal value.
      */
-    public function get($key)
+    public function get($key, $default = null)
     {
+        $this->validateKey($key);
+
         if (isset($this->cache[$key])) {
             return $this->cache[$key]->data();
         }
 
-        return null;
+        return $default;
     }
 
     /**
-     * $ttl is ignored. If your data is living that long in memory, you got issues.
+     * Persists data in the cache, uniquely referenced by a key with an optional expiration TTL time.
      *
-     * {@inheritdoc}
+     * @param string $key The key of the item to store.
+     * @param mixed $value The value of the item to store, must be serializable.
+     * @param null|int|DateInterval $ttl Optional. The TTL value of this item. If no value is sent and
+     *                                     the driver supports TTL then the library may set a default value
+     *                                     for it or let the driver take care of that.
+     *
+     * @return bool True on success and false on failure.
+     *
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     *   MUST be thrown if the $key string is not a legal value.
      */
-    public function set($key, $value, $ttl = 0)
+    public function set($key, $value, $ttl = null)
     {
+        $this->validateKey($key);
+
         $this->cache[$key] = new Item($value);
         return true;
+    }
+
+    /**
+     * Delete an item from the cache by its unique key.
+     *
+     * @param string $key The unique cache key of the item to delete.
+     *
+     * @return bool True if the item was successfully removed. False if there was an error.
+     *
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     *   MUST be thrown if the $key string is not a legal value.
+     */
+    public function delete($key)
+    {
+        $this->validateKey($key);
+
+        unset($this->cache[$key]);
+        return true;
+    }
+
+    /**
+     * Wipes clean the entire cache's keys.
+     *
+     * @return bool True on success and false on failure.
+     */
+    public function clear()
+    {
+        $this->cache = [];
+        return true;
+    }
+
+    /**
+     * Obtains multiple cache items by their unique keys.
+     *
+     * @param iterable $keys A list of keys that can obtained in a single operation.
+     * @param mixed $default Default value to return for keys that do not exist.
+     *
+     * @return iterable A list of key => value pairs. Cache keys that do not exist or are stale will have $default as value.
+     *
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     *   MUST be thrown if $keys is neither an array nor a Traversable,
+     *   or if any of the $keys are not a legal value.
+     */
+    public function getMultiple($keys, $default = null)
+    {
+        $this->validateIterable($keys);
+
+        $cacheData = [];
+        foreach ($keys as $key) {
+            $cacheData[$key] = $this->get($key, $default);
+        }
+
+        return $cacheData;
+    }
+
+    /**
+     * Persists a set of key => value pairs in the cache, with an optional TTL.
+     *
+     * @param iterable $values A list of key => value pairs for a multiple-set operation.
+     * @param null|int|DateInterval $ttl Optional. The TTL value of this item. If no value is sent and
+     *                                      the driver supports TTL then the library may set a default value
+     *                                      for it or let the driver take care of that.
+     *
+     * @return bool True on success and false on failure.
+     *
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     *   MUST be thrown if $values is neither an array nor a Traversable,
+     *   or if any of the $values are not a legal value.
+     */
+    public function setMultiple($values, $ttl = null)
+    {
+        $this->validateIterable($values);
+
+        foreach ($values as $key => $value) {
+            $this->set($key, $value);
+        }
+
+        return true;
+    }
+
+    /**
+     * Deletes multiple cache items in a single operation.
+     *
+     * @param iterable $keys A list of string-based keys to be deleted.
+     *
+     * @return bool True if the items were successfully removed. False if there was an error.
+     *
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     *   MUST be thrown if $keys is neither an array nor a Traversable,
+     *   or if any of the $keys are not a legal value.
+     */
+    public function deleteMultiple($keys)
+    {
+        $this->validateIterable($keys);
+
+        foreach ($keys as $key) {
+            $this->delete($key);
+        }
+
+        return true;
+    }
+
+    /**
+     * Determines whether an item is present in the cache.
+     *
+     * NOTE: It is recommended that has() is only to be used for cache warming type purposes
+     * and not to be used within your live applications operations for get/set, as this method
+     * is subject to a race condition where your has() will return true and immediately after,
+     * another script can remove it making the state of your app out of date.
+     *
+     * @param string $key The cache item key.
+     *
+     * @return bool
+     *
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     *   MUST be thrown if the $key string is not a legal value.
+     */
+    public function has($key)
+    {
+        $this->validateKey($key);
+
+        return isset($this->cache[$key]);
     }
 }